refactor structure a bit and fix some minor issues
